third_party/blink/web_tests/external/wpt/html/semantics/scripting-1/the-script-element/text-module/credentials.sub-expected.txt

This is a testharness.js-based test.
[FAIL] Text Modules should be loaded with or without the credentials based on the same-origin-ness and the crossOrigin attribute
  assert_equals: Descendant text modules should be loaded with the credentials when the crossOrigin attribute is not specified and the target is same-origin expected (string) "request had cookies" but got (undefined) undefined
Harness: the test ran to completion.
